my-ai-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
pydantic[email]==1.10.13
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
itsdangerous==2.1.2
python-dateutil==2.9.0.post0
